refactor: update paths to SVG icons
